Bug Fix, Endless menu changes
FIX
-An error would be thrown when couldn't attack to check if the move was an ultimate

NEW
-Can see how much gold was gained on the last round

ADJUSTS
-Rearenged the endless menu

+++ Jogo/Game Project/Assets/Code/EndlessData.cs
@@ -24,7 +24,7 @@ public class EndlessData
     public float playerSan;
     public float playerUlt;
 
-    public float gold;
+    public int gold;
     public int level;
     public int round;
